+++ apps/chat/src/store/publication/publication.types.ts
@@ -38,7 +38,7 @@ export interface PublicationState {
   isEditMode: boolean;
   entitiesEditState: Record<string, { name: string; version: string }>;
   foldersEditState: FolderEditTree;
-  rulesOnEdit?: PublicationRule[];
+  rulesOnEdit: PublicationRule[];
   isPublicationUpdating: boolean;
   displayAuthorEditState: string;
 }
